generate-assets/requirements.txt
mutagen
google-cloud-texttospeech
autocorrect
playsound
jamspell
nltk